feat: Conditionally display hourly forecast
This commit introduces a change to the table output format. The "Hourly Forecast" section will now only be displayed if the `forecast-hours` configuration option is greater than 0.

This improves the user experience by hiding the hourly forecast section when it is not relevant.

The unit tests have been updated to reflect this change.